Moved Category, Session to "website" app. Will afterwards use "website" as common app.
