+++ src/problem/admin.py
@@ -2,11 +2,6 @@ from django.contrib import admin
 from .models import *
 
 
-@admin.register(ProblemCategory)
-class ProblemCategoryAdmin(admin.ModelAdmin):
-    list_display = ('title', 'description')
-
-
 @admin.register(Problem)
 class ProblemAdmin(admin.ModelAdmin):
     list_display = ('title', 'categories_name', 'author', 'description')
