+++ src/seminar/templates/seminar/list.html
