+++ src/website/models.py
@@ -11,3 +11,26 @@ class User(AbstractUser):
         verbose_name_plural = '사용자들'
 
 
+class Category(models.Model):
+    title = models.CharField(max_length=30, unique=True)
+    description = models.TextField(blank=True)
+
+    class Meta:
+        verbose_name = '카테고리'
+        verbose_name_plural = '카테고리들'
+
+    def __str__(self):
+        return '%s' % self.title
+
+
+class Session(models.Model):
+    title = models.CharField(max_length=50, unique=True)
+    description = models.TextField(blank=True)
+    isActive = models.BooleanField(default=False)
+
+    class Meta:
+        verbose_name = '분기'
+        verbose_name_plural = '분기들'
+
+    def __str__(self):
+        return '%s' % self.title
